requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
